Обучение по лучшим практикам: QLoRA-пайплайн + рефакторинг тренера под M4 16GB
Путь 1 (рекомендуемый): QLoRA-дообучение Gemma 3 4B (4-bit QAT)
- finetune/build_chat_dataset.py: русский чат-датасет из saiga/grandmaster/dialogues
  с фильтрами качества, дедупликацией и persona-миксом (5724 train / 276 valid)
- finetune/qlora_config.yaml: mask_prompt, LoRA на все слои (rank 16),
  cosine LR + warmup, grad checkpointing, grad accumulation
- train_qlora.sh / chat.sh / fuse.sh обновлены

Путь 2: обучение с нуля (tasochka)
- model_mlx/model_torch: GQA (num_kv_heads), tied embeddings, init 0.02/sqrt(2L),
  fused causal-маска SDPA; обратная совместимость со старыми чекпоинтами
- train_mlx: mx.compile, --bf16 (~2x скорость), --grad-accum, AdamW betas (0.9, 0.95),
  wd 0.1, точный resume (оптимизатор + шаг), best-val чекпоинт, tokens/sec
- train_torch: паритет (betas, раздельный weight decay, resume, best-val)
- generate_mlx: фикс генерации с bf16-весами

Docs: TRAINING_GUIDE.md с готовыми командами и таблицами «что менять»

+++ tasochka/generate_mlx.py
@@ -66,7 +66,8 @@ def _sample_token(
     banned_ids: Optional[Set[int]] = None,
 ) -> int:
     mx.eval(logits)
-    logits_np: np.ndarray = np.array(logits).astype(np.float64)
+    # fp32 cast first: numpy can't read bf16 buffers directly
+    logits_np: np.ndarray = np.array(logits.astype(mx.float32)).astype(np.float64)
 
     # 1. Repetition penalty: pull recent tokens toward zero
     if rep_penalty != 1.0 and recent_ids:
